+++ src/factories/digi/SiliconPulseGeneration_factory.h
@@ -25,9 +25,9 @@ private:
 
   ParameterRef<std::string> m_pulse_shape_function{this, "pulseShapeFunction", config().pulse_shape_function};
   ParameterRef<std::vector<double>> m_pulse_shape_params{this, "pulseShapeParams", config().pulse_shape_params};
-  ParameterRef<float> m_timestep{this, "timestep", config().timestep};
-  ParameterRef<float> m_ignore_thres{this, "ignoreThreshold", config().ignore_thres};
-  ParameterRef<float> m_min_sampling_time{this, "minSamplingTime", config().min_sampling_time};
+  ParameterRef<double> m_timestep{this, "timestep", config().timestep};
+  ParameterRef<double> m_ignore_thres{this, "ignoreThreshold", config().ignore_thres};
+  ParameterRef<double> m_min_sampling_time{this, "minSamplingTime", config().min_sampling_time};
   ParameterRef<int> m_max_time_bins{this, "maxTimeBins", config().max_time_bins};
 
   Service<AlgorithmsInit_service> m_algorithmsInit{this};
